+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImp.java
@@ -1,19 +1,14 @@
 package ru.kata.spring.boot_security.demo.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.kata.spring.boot_security.demo.dao.UserDao;
-import ru.kata.spring.boot_security.demo.model.Role;
 import ru.kata.spring.boot_security.demo.model.User;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Service
 public class UserServiceImp implements UserService {
@@ -40,7 +35,7 @@ public class UserServiceImp implements UserService {
 
     @Transactional(readOnly = true)
     @Override
-    public User get(int id) {
+    public User get(Long id) {
         return dao.get(id);
     }
 
@@ -63,7 +58,7 @@ public class UserServiceImp implements UserService {
 
     @Transactional
     @Override
-    public void delete(int id) {
+    public void delete(Long id) {
         dao.delete(id);
     }
